refactor initPodData function from podDataSlice
